% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_composantes_principales.R
\name{analyse_composantes_principales}
\alias{analyse_composantes_principales}
\title{Analyse en Composantes Principales (ACP) et génération d'un biplot}
\usage{
analyse_composantes_principales(donnees)
}
\arguments{
\item{donnees}{Un data.frame contenant les données pour lesquelles l'ACP est effectuée.
Les colonnes doivent être des variables quantitatives sur lesquelles l'ACP sera réalisée.}
}
\value{
Un objet de type `ggplot` représentant le biplot des deux premières composantes principales.
  Le biplot permet de visualiser la relation entre les observations (points) et les contributions des variables (flèches).
}
\description{
Cette fonction effectue une analyse en composantes principales (ACP) sur un jeu de données
et génère un biplot pour visualiser les projections des observations et les contributions des variables.
}
\details{
Cette fonction applique l'Analyse en Composantes Principales (ACP) à un jeu de données. Elle effectue d'abord une standardisation des variables (centrage et réduction), puis génère un biplot qui projette les observations dans l'espace des premières composantes principales.
Le biplot affiche également les contributions des variables sous forme de flèches pointant dans la direction de chaque composante principale.

La taille des flèches est automatiquement ajustée pour garantir leur visibilité. Les observations sont représentées par des points, et les flèches sont colorées en rouge avec les noms des variables associés à chaque flèche.
}
\note{
Si le jeu de données contient des valeurs manquantes (`NA`), celles-ci seront automatiquement supprimées avant de réaliser l'ACP.
}
\examples{
# Exemple d'utilisation de la fonction analyse_composantes_principales
donnees <- data.frame(
  actif_1 = c(0.01, 0.02, -0.01, 0.03),
  actif_2 = c(0.03, 0.02, 0.01, 0.02),
  actif_3 = c(0.04, -0.03, 0.02, 0.05)
)
analyse_composantes_principales(donnees)

}
