% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrice_covariance.R
\name{matrice_covariance}
\alias{matrice_covariance}
\title{Calcul de la matrice de covariance}
\usage{
matrice_covariance(donnees)
}
\arguments{
\item{donnees}{Un data.frame contenant les données pour lesquelles la matrice de covariance
doit être calculée.}
}
\value{
Une matrice de covariance.
}
\description{
Cette fonction calcule la matrice de covariance à partir d'un jeu de données.
}
\examples{
# Exemple d'utilisation de la fonction matrice_covariance
donnees <- data.frame(
  actif_1 = c(0.01, 0.02, -0.01, 0.03),
  actif_2 = c(0.03, 0.02, 0.01, 0.02)
)
matrice_covariance(donnees)  # Appel de la fonction matrice_covariance

}
