% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele_risque.R
\name{modele_risque}
\alias{modele_risque}
\title{Modélise le risque avec un modèle factoriel}
\usage{
modele_risque(donnees, facteurs)
}
\arguments{
\item{donnees}{Un data.frame contenant les rendements des actifs.}

\item{facteurs}{Un data.frame contenant les facteurs de risque.}
}
\value{
Un vecteur de coefficients représentant la sensibilité des actifs
aux facteurs de risque.
}
\description{
Cette fonction utilise un modèle de régression linéaire pour calculer les
coefficients de sensibilité des actifs par rapport à des facteurs de risque.
}
\examples{
# Exemple de données fictives pour les actifs et les facteurs de risque
donnees <- data.frame(actif_1 = rnorm(100, mean = 0.05, sd = 0.2),  # Actif 1 : rendements simulés
  actif_2 = rnorm(100, mean = 0.03, sd = 0.15), # Actif 2 : rendements simulés
actif_3 = rnorm(100, mean = 0.04, sd = 0.1),  # Actif 3 : rendements simulés
actif_4 = rnorm(100, mean = 0.06, sd = 0.25), # Actif 4 : rendements simulés
actif_5 = rnorm(100, mean = 0.02, sd = 0.18)  # Actif 5 : rendements simulés
)

facteurs <- data.frame(facteur_1 = rnorm(100, mean = 0.02, sd = 0.1), # Facteur 1 : exemple
                      facteur_2 = rnorm(100, mean = 0.03, sd = 0.08)  # Facteur 2 : exemple
 )

modele_risque(donnees, facteurs)

}
