% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_portefeuille.R
\name{optim_portefeuille}
\alias{optim_portefeuille}
\title{Optimisation du portefeuille}
\usage{
optim_portefeuille(covariance, rendements, niveau_risque)
}
\arguments{
\item{covariance}{Une matrice de covariance des rendements des actifs.}

\item{rendements}{Un vecteur des rendements attendus des actifs.}

\item{niveau_risque}{Un nombre représentant le niveau de risque (la variance cible) du portefeuille.}
}
\value{
Un vecteur contenant les pondérations optimales des actifs dans le portefeuille.
}
\description{
Cette fonction optimise un portefeuille en utilisant une fonction objectif quadratique,
avec comme paramètres la matrice de covariance des rendements des actifs, les rendements attendus
et un niveau de risque cible.
}
\examples{
# Exemple d'utilisation
covariance <- matrix(c(0.04, 0.02, 0.02, 0.03), nrow = 2)  # Matrice de covariance des actifs
rendements <- c(0.06, 0.08)  # Rendements attendus des actifs
niveau_risque <- 0.005  # Niveau de risque cible
optim_portefeuille(covariance, rendements, niveau_risque)

}
